requirements/base.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Django==4.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
idna==3.4
mypy-extensions==0.4.3
pathspec==0.10.1
pip-autoremove==0.10.0
pipdeptree==2.3.1
platformdirs==2.5.2
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.5.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
types-cryptography==3.3.23
typing_extensions==4.1.1
urllib3==1.26.12
